requirements.txt
asgiref==3.10.0
charset-normalizer==3.4.4
Django==5.2.8
django-filter==25.2
djangorestframework==3.16.1
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==12.0.0
psycopg2-binary==2.9.11
reportlab==4.4.4
sqlparse==0.5.3
tzdata==2025.2
